feat : Q4 (computeVoteSummary) done

+++ src/main/java/main/CandidateVoteCount.java
@@ -10,5 +10,5 @@ import lombok.*;
 
 public class CandidateVoteCount {
     private String candidateName;
-    private int ValidVoteCount;
+    private long ValidVoteCount;
 }
